v1.01
DB_Controller renamed to DataBaseProvider due to it is not a controller
All the models is now extend Base_Model class. It is contains DataBaseProvider and AjaxResponder members in
Models requirement is now two-level folder depth.
All DataBaseProvider methods (except setup method) are not static anymore. You can get access to it functions via creating of DataBaseProvider object or from Controller ($this->DataBaseProvider->MethodName()).
snake_case switched to CamelCase